requirements.txt
APScheduler==3.6.3
bleak==0.9.1
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
decorator==4.4.2
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.19.4
opencv-contrib-python==4.4.0.44
opencv-python==4.4.0.46
paho-mqtt==1.5.1
pycparser==2.20
python-telegram-bot==11.1.0
pytz==2020.4
six==1.15.0
tornado==6.1
tzlocal==2.1
Werkzeug==1.0.1
